requirements.txt
Flask
authlib
python-dotenv
pytz
requests
